requirements.txt
flask
vercel
numpy
psycopg2-binary

